ADD: app error

+++ src/storage/group/groupCreate.ts
@@ -1,10 +1,17 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
+import { AppError } from "@utils/AppError";
 import { GROUP_COLLECTION } from "../storageConfig";
 import { GroupsGetAll } from "./groupsGetAll";
 
 export async function groupCreate(newGroupName: string) {
   try {
     const storedGroups = await GroupsGetAll();
+    const groupAlreadyExists = storedGroups.includes(newGroupName);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Já existe um grupo com o nome escolhido!')
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroupName]);
 
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
